refactor(auth): move auth controller to correct directory structure
Update imports in routes and tests to reflect new controller location
Add new auth controller implementation with full feature set
